Update README.md
Added the heading "Train Model" because of the separation of training the model and solving power flow in the "solve_power_flow.py" script. Also changed "Data Generation" to "Data Generation Process" to avoid using "Data Generation Folder" as a heading. Included the remaining information and added a "Dependencies" heading for completion.